Move the benchmark with related benchmarks
I don't know why a bunch of samplers-package related tests are in
that root file...
