+++ parser_test.go
@@ -1110,63 +1110,3 @@ func BenchmarkParseMetric(b *testing.B) {
 	// Attempt to avoid compiler optimizations? Is this relevant?
 	benchResult = total
 }
-
-var benchResultInt int
-
-func BenchmarkConvertIndicatorMetrics(b *testing.B) {
-	explicitTags := []struct {
-		name       string
-		metricTags string
-	}{
-		{"ZeroTags", ""},
-		{"OneTag", "baz:gorch"},
-		{"TwoTags", "foo:bar,baz:gorch"},
-	}
-	implicitTags := []struct {
-		name       string
-		metricTags []string
-	}{
-		{"NoImplicit", []string{}},
-		{"OneImplicit", []string{"six:6"}},
-		{"SixImplicit", []string{"three:three", "six:6", "nine:9", "ten:10", "eleven:11", "twelve:12"}},
-	}
-
-	total := 0
-
-	for _, it := range implicitTags {
-		for _, et := range explicitTags {
-			duration := 5 * time.Second
-			start := time.Now()
-			end := start.Add(duration)
-			span := &ssf.SSFSpan{}
-			span.Id = 1
-			span.TraceId = 5
-			span.Name = "foo"
-			span.StartTimestamp = start.UnixNano()
-			span.EndTimestamp = end.UnixNano()
-			span.Indicator = true
-			span.Service = "bar-srv"
-			span.Metrics = make([]*ssf.SSFSample, 0)
-			buff, err := proto.Marshal(span)
-			assert.Nil(b, err)
-			inSpan, err := protocol.ParseSSF(buff)
-			if err != nil {
-				b.Fatal(err)
-			}
-
-			benchName := fmt.Sprintf("%s%s", it.name, et.name)
-			parser := samplers.NewParser(it.metricTags)
-			b.Run(benchName, func(b *testing.B) {
-				for n := 0; n < b.N; n++ {
-					ums, err := parser.ConvertIndicatorMetrics(inSpan, "", "timer_name")
-					if err != nil {
-						b.Fatal(err)
-					}
-					total += len(ums)
-				}
-			})
-		}
-	}
-	// Attempt to avoid compiler optimizations? Is this relevant?
-	benchResultInt = total
-}
